#376: refactor
